backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
passlib
python-jose[cryptography]
python-multipart
email-validator